Refactor: Refactoring scripts
