--- aggregate_time.csv
@@ -1,2 +1,0 @@
-set_time,set_calls_count,aggregate_set_time,get_time,get_calls_count,aggregate_get_time,show_all_time,show_all_calls_count,aggregate_show_all_time
-52.240761518478394,29556,0.001767517983437488,96.09460234642029,49698,0.001933570814648885,0.030834436416625977,3,0.010278145472208658
